🚧 (Models) : Updated getters and setters
