+++ apps/api/src/main/java/com/fridgetoplate/controller/RecommendController.java
@@ -3,8 +3,6 @@ package com.fridgetoplate.controller;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
-import com.fridgetoplate.interfaces.Recipe;
 import com.fridgetoplate.model.RecipeModel;
 import com.fridgetoplate.repository.RecipeRepository;
